+++ src/qililab/experiment/portfolio/flipping_sequence.py
@@ -46,7 +46,7 @@ class FlippingSequence(ExperimentAnalysis, Cos):
             circuit.add(M(qubit))
             circuits.append(circuit)
 
-        control_bus, readout_bus = platform.get_bus_by_qubit_index(qubit)
+        _, control_bus, readout_bus = platform.get_bus_by_qubit_index(qubit)
 
         experiment_options = ExperimentOptions(
             name="Flipping Sequence",