+++ src/qililab/instruments/awg_settings/awg_qblox_sequencer.py
@@ -11,9 +11,7 @@ class AWGQbloxSequencer(AWGSequencer):
     """AWG Qblox Sequencer
 
     Args:
-        sync_enabled (bool): Enable synchronization over multiple instruments for this sequencer.
         num_bins (int): Number of bins
     """
 
-    sync_enabled: bool
     num_bins: int